tests added modules implemented
