--- SecondLesson/CarRent/CarRentTests/UnitTest1.cs
@@ -1,13 +1,0 @@
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace CarRentTests
-{
-    [TestClass]
-    public class UnitTest1
-    {
-        [TestMethod]
-        public void TestMethod1()
-        {
-        }
-    }
-}
