+++ SecondLesson/CarRent/EmptyClass.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Application
+{
+    public class EmptyClass
+    {
+        public EmptyClass()
+        {
+        }
+    }
+}
